backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pandas==2.3.2
openpyxl==3.1.5
boto3==1.40.22
python-dotenv==1.1.1
